+++ src/ChannelService.php
@@ -9,6 +9,8 @@ use StadGent\Services\OpeningHours\Request\Channel\GetAllRequest;
 use StadGent\Services\OpeningHours\Request\Channel\GetByIdRequest;
 use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursDayHtmlRequest;
 use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursDayRequest;
+use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursWeekHtmlRequest;
+use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursWeekRequest;
 use StadGent\Services\OpeningHours\Request\Channel\OpenNowHtmlRequest;
 use StadGent\Services\OpeningHours\Request\Channel\OpenNowRequest;
 use StadGent\Services\OpeningHours\Request\RequestInterface;
@@ -267,6 +269,69 @@ class ChannelService extends ServiceAbstract implements CacheableInterface
         return $this->sendHtmlRequest(
             $cacheKey,
             new OpeningHoursDayHtmlRequest($serviceId, $channelId, $date)
+        );
+    }
+
+    /**
+     * Get the Opening Hours for a single week as Value object.
+     *
+     * @param int $serviceId
+     *   The Service ID.
+     * @param int $channelId
+     *   The Channel ID.
+     * @param string $date
+     *   The start date (Y-m-d) of the week period to get the data for.
+     *
+     * @return \StadGent\Services\OpeningHours\Value\OpeningHours
+     *
+     * @throws \Exception
+     * @throws \GuzzleHttp\Exception\RequestException
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     * @throws \StadGent\Services\OpeningHours\Exception\NotFoundException
+     * @throws \StadGent\Services\OpeningHours\Exception\ChannelNotFoundException
+     * @throws \StadGent\Services\OpeningHours\Exception\ServiceNotFoundException
+     */
+    public function openingHoursWeek($serviceId, $channelId, $date)
+    {
+        $cacheKey = $this->createCacheKeyFromArray(
+            [__FUNCTION__, $serviceId, $channelId, $date]
+        );
+
+        return $this->sendOpeninghoursRequest(
+            $cacheKey,
+            new OpeningHoursWeekRequest($serviceId, $channelId, $date)
+        );
+    }
+
+    /**
+     * Get the Opening Hours for a single week as HTML.
+     *
+     * @param int $serviceId
+     *   The Service ID.
+     * @param int $channelId
+     *   The Channel ID.
+     * @param string $date
+     *   The start date (Y-m-d) of the week period to get the data for.
+     *
+     * @return string
+     *   The HTML.
+     *
+     * @throws \Exception
+     * @throws \GuzzleHttp\Exception\RequestException
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     * @throws \StadGent\Services\OpeningHours\Exception\NotFoundException
+     * @throws \StadGent\Services\OpeningHours\Exception\ChannelNotFoundException
+     * @throws \StadGent\Services\OpeningHours\Exception\ServiceNotFoundException
+     */
+    public function openingHoursWeekHtml($serviceId, $channelId, $date)
+    {
+        $cacheKey = $this->createCacheKeyFromArray(
+            [__FUNCTION__, $serviceId, $channelId, $date]
+        );
+
+        return $this->sendHtmlRequest(
+            $cacheKey,
+            new OpeningHoursWeekHtmlRequest($serviceId, $channelId, $date)
         );
     }