+++ CHANGELOG.md
@@ -16,5 +16,7 @@ All Notable changes to `gent/services-opening-hours` package.
 * DMOH-12: Added the OpenNow & OpenNowHTML methods to the ChannelService.
 * DMOH-13: Added the OpeningHoursDay & OpeningHoursDayHTML methods to the
   ChannelService.
+* DMOH-14: Added the OpeningHoursWeek & OpeningHoursWeekHTML methods to the
+  ChannelService.
 
 [Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop